fix: root level does not follow global level
When max_log_level is configured to more verbose than `INFO`, users find
no log other than `INFO` or above is printed.
This is due to `root_log_level` sets a default to `INFO`, which is not
affected by `max_log_level`.
This is behaviour is anti-intuitive.

With this fix, `root_log_level` will first check whether it is set, and
otherwises use `max_log_level` as default.